Use question_ids instead of field.id for matching
Duplicated Typeform Forms do not use the field_id as a canonical
ID for finding answers to questions, and instead rely on matching
the question_id.

This commit alters the logic for finding matching fields to check
against inclusion in the question_ids array, which should more
reliably work both for duplicated and non-duplicated forms.